requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.13.3
black==25.1.0
cchardet==2.1.7
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.115.11
frozenlist==1.5.0
h11==0.14.0
idna==3.10
importlib_metadata==8.6.1
multidict==6.1.0
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==24.2
pathspec==0.12.1
pefile==2023.2.7
pillow==11.1.0
platformdirs==4.3.6
propcache==0.3.0
pycares==4.5.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-whois==0.9.5
pythonping==1.1.4
pywin32-ctypes==0.2.3
requests==2.32.3
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.46.1
tomli==2.2.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
websocket-client==1.8.0
winloop==0.1.8
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
